Rename replace-plugin to helpers
